+++ CHANGELOG.md
@@ -27,6 +27,7 @@ That said, these are more guidelines rather than hardset rules, though the proje
 - [#1830](https://github.com/ClementTsang/bottom/pull/1830): Add spacebar shortcut to toggle process tree expansion.
 - [#1861](https://github.com/ClementTsang/bottom/pull/1861): Add read-only mode, where things like killing processes is disabled.
 - [#1890](https://github.com/ClementTsang/bottom/pull/1890): Add enter key shortcut to close process search widget.
+- Support native Intel GPU monitoring via sysfs on Linux.
 
 ### Bug Fixes